+++ server/src/network/session/mod.rs
@@ -1,19 +1,6 @@
-pub use compound_collector::*;
-pub use order_channel::*;
-pub use receive::*;
-pub use recovery_queue::*;
-pub use send::*;
-pub use send_queue::*;
-pub use session::*;
-pub use session_manager::*;
+use common::glob_export;
 
-mod compound_collector;
-mod order_channel;
-mod receive;
-mod recovery_queue;
-mod send;
-mod send_queue;
-mod session;
-mod session_manager;
+glob_export!(session);
+glob_export!(manager);
 
 pub mod handlers;
