+++ server/src/test.rs
@@ -7,8 +7,8 @@ use flate2::read::DeflateDecoder;
 use tokio::net::windows::named_pipe::PipeMode::Byte;
 
 use crate::instance_manager::{IPV4_LOCAL_ADDR, IPV6_LOCAL_ADDR};
-use crate::network::raknet::{Frame, Header};
-use crate::network::session::OrderChannel;
+use crate::network::Header;
+use crate::network::raknet::{Frame, OrderChannel};
 use common::VResult;
 use common::{ReadExtensions, WriteExtensions};
